# Conflicts:
#	tests/feature_selection/biz_val/test_biz_val_target_types_coverage.py
#	tests/feature_selection/contracts/test_integration_contract.py
#	tests/feature_selection/fe/categorical/test_cat_fe_tier4_real.py
#	tests/feature_selection/mrmr/biz_val/test_biz_val_filters_mrmr_mi_correction.py
#	tests/feature_selection/mrmr/core/test_mrmr_edge_cases.py
#	tests/feature_selection/mrmr/core/test_mrmr_screen_discretization_fixes.py
#	tests/feature_selection/mrmr/core/test_mrmr_sklearn_contract_and_threadlocals.py
#	tests/feature_selection/mrmr/core/test_mrmr_synergy_redundancy_adversarial.py
#	tests/feature_selection/mrmr/fe/test_biz_val_mrmr_sample_weight_fe.py
#	tests/feature_selection/robustness/test_edge_cases_robustness.py
#	tests/feature_selection/test_fs_coverage_gaps.py
#	tests/feature_selection/test_fs_defaults_and_fallback_guards.py
#	tests/feature_selection/test_fs_groups_recipe_seed_fixes.py

+++ tests/feature_engineering/test_graph_spectral_features.py
@@ -100,7 +100,7 @@ def _matched_graph_of_class(cls, n, m, rng):
     else:
         within = [(i, j) for (i, j) in all_pairs if (i < half) == (j < half)]
         between = [(i, j) for (i, j) in all_pairs if (i < half) != (j < half)]
-        n_within = min(int(round(0.9 * m)), len(within))
+        n_within = min(round(0.9 * m), len(within))
         n_between = min(m - n_within, len(between))
         wi = rng.choice(len(within), size=n_within, replace=False)
         bi = rng.choice(len(between), size=n_between, replace=False)
